src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-annotate pyproject.toml -o requirements.txt
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
ghp-import==2.1.0
idna==3.10
jinja2==3.1.4
markdown==3.7
markupsafe==3.0.1
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.41
mkdocs-material-extensions==1.3.1
packaging==24.1
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.6
pygments==2.18.0
pymdown-extensions==10.11.2
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pyyaml-env-tag==0.1
regex==2024.9.11
requests==2.32.3
six==1.16.0
urllib3==2.2.3
watchdog==5.0.3
